+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Git (3.0)
+- Add `PackedSeqVec::clear()`.
 - `PackedSeq::as_u64` now supports k up to 32, instead of only up to 29. (Fixes #2.)
 - `PackedSeqVec::get` now uses checked indexing, to prevent unsafe out-of-bounds access.
 - **Breaking**: `PackedSeqVec::{seq, len}` are now private, to uphold internal